passkey: propagate error code for power cycle
When performing a passkey authentication and the PIN is entered 3 times
incorrectly the FIDO2 device requires a power cycle (disconnect and
reconnect the device to the USB port). libfido2 recognizes this with a
special error code and the passkey child should return it so that the
SSSD responder is aware of it.

Signed-off-by: Iker Pedrosa <ipedrosa@redhat.com>
(cherry picked from commit 6754892df19c6d1767480494461c9a3a38017cb8)