ci: add autobackport workflow
This adds an autobackport workflow that will automatically backport
pull request with backport-to-$branch label. Backport pull request
is created upon mergin the original pull request for each requested
branch. It is also possible to create additional backport by labeling
the already merged pull request.

If there are any conflicts during cherry-picks, they are commited and
pull request is still opened. The author of the original pull request
can then fix them and push to the backport pull request (guide is
presented in the PR description).