ci: automatically add Reviewed-by trailer when Accepted label is set
This introduce new workflow that is run when Accepted label is set.
It appends Reviewed-by trailers to all commits in the pull request for
all reviewers that ack-ed it.

* It is skipped if Accepted label is missing/removed
* It is no-op success if PR was synchronized after label was set (this
  is required to make it green as it is automatically re-run when the
  job finishes as a side effect to pushing to this repository)
* It adds the trailers when Accepted label is added

It uses secrets.BOT_TOKEN and SSSD_AUTHORS variable that contains
mapping from github account to name and email.

Since PR CI is restarted when commits with Reviewed-by are pushed, the
job also submits comment with current PR CI status, so it is not lost
and we do not have to necessarily wait for new PR CI run to finish.